+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Next
+# 0.1.0
 
 - Removed "type" field from the unique team list
 - Added missing localizations